Added User Profile and user has single role

+++ src/main/java/me/abulbasar/bulksms/controller/PackageController.java
@@ -10,6 +10,7 @@ import java.util.List;
 import me.abulbasar.bulksms.exception.ResourceNotFoundException;
 import me.abulbasar.bulksms.model.Package;
 import me.abulbasar.bulksms.repository.PackageRepository;
+import me.abulbasar.bulksms.service.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,9 +19,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import org.springframework.web.util.UriComponentsBuilder;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 /**
  *
@@ -32,63 +31,91 @@ public class PackageController {
 
     @Autowired
     PackageRepository packageRepository;
+    
+    @Autowired
+    SecurityService securityService;
+
 
     @GetMapping("/list")
-    public String listPacage(Model model, @RequestParam(required=false) String sm) {
+    public String listPacage(Model model,
+                             @PathVariable("role") String role) {
         List<Package> packages = packageRepository.findAll();
         model.addAttribute("packages", packages);
-        model.addAttribute("urlRole", getUrlRole());
-        if(sm != null) {
-            model.addAttribute("sm", sm);
-        }
+        model.addAttribute("role", role);
+
         return "packageList";
     }
 
     @GetMapping("/add")
-    public String addPage(Model model) {
+    public String addPage(Model model, @PathVariable("role") String role) {
         Package pack = new Package();
         model.addAttribute("pack", pack);
-        model.addAttribute("urlRole", getUrlRole());
+        model.addAttribute("role", role);
+
         return "packageAdd";
     }
 
     @PostMapping("/add")
-    public String addPackage(@ModelAttribute Package pack) {
+    public String addPackage(@ModelAttribute Package pack,
+                             RedirectAttributes redirectAttributes) {
         packageRepository.save(pack);
-        return "redirect:list?sm=Add Successful";
+        redirectAttributes.addFlashAttribute("sm", "Package Add Successful");
+        
+        return "redirect:list";
     }
-    
+
     @GetMapping("/update/{id}")
-    public String updatePage(@PathVariable("id") int id, Model model, @RequestParam(required=false) String sm) {
-        Package pack = packageRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Package", "id", id));
+    public String updatePage(Model model,
+                             @PathVariable("id") int id,
+                             @PathVariable("role") String role) {
+        Package pack = packageRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Package", "id", id));
         model.addAttribute("pack", pack);
-        model.addAttribute("urlRole", getUrlRole());
-        if(sm != null) {
-            model.addAttribute("sm", sm);
-        }
+        model.addAttribute("role", role);
+
         return "packageUpdate";
     }
-    
+
     @PostMapping("/update")
-    public String updatePackage(@ModelAttribute Package newPack) {
-        Package pack = packageRepository.findById(newPack.getId()).orElseThrow(()-> new ResourceNotFoundException("Package", "id", newPack.getId()));
+    public String updatePackage(@ModelAttribute Package newPack,
+                                RedirectAttributes redirectAttributes) {
+        if(!securityService.findUserRole().equals("ADMIN")) {
+            redirectAttributes.addFlashAttribute("em", "Unauthorized Access");
+            return "redirect:list";
+        }
+        
+        Package pack = packageRepository.findById(newPack.getId()).orElseThrow(
+                () -> new ResourceNotFoundException("Package", "id",
+                                                    newPack.getId()));
         pack.setpName(newPack.getpName());
         pack.setPrice(newPack.getPrice());
         pack.setSms(newPack.getSms());
         pack.setDuration(newPack.getDuration());
         pack.setType(newPack.getType());
+        pack.setActivated(newPack.isActivated());
         packageRepository.save(pack);
-        return "redirect:update/"+pack.getId()+"?sm=Update Succesful";
+        redirectAttributes.addFlashAttribute("sm", "Package Update Successful");
+        
+        return "redirect:update/"+pack.getId();
     }
-    
+
     @GetMapping("/delete/{id}")
-    public String deletePackage(@PathVariable("id") int id) {
-        Package pack = packageRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Package", "id", id));
+    public String deletePackage(@PathVariable("id") int id,
+                                RedirectAttributes redirectAttributes) {
+        System.out.println("Enter Delete Package Controller ...\n");
+        if(!securityService.findUserRole().equals("ADMIN")) {
+            redirectAttributes.addFlashAttribute("em", "Unauthorized Access");
+            return "redirect:../list";
+        }
+
+        Package pack = packageRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Package", "id", id));
         packageRepository.delete(pack);
-        return "redirect:../list?sm=Delete Succesful";
+        System.out.println("Package Delete Success...\n");
+        redirectAttributes.addFlashAttribute("em", "Package Deleted");
+        System.out.println("Redirecting to the list with sm....\n\n");
+        return "redirect:../list";
     }
-    
-    
 
     @ModelAttribute("types")
     public List<String> getTypes() {
@@ -98,15 +125,4 @@ public class PackageController {
         return types;
     }
 
-    String getUrlRole() {
-        UriComponentsBuilder builder = ServletUriComponentsBuilder.fromCurrentRequest();
-        String fullPath = builder.buildAndExpand().getPath();
-        String target = "/bulksms/";
-        String afterContext = fullPath.substring(target.length());
-        String role = afterContext;
-        if(afterContext.contains("/")) {
-            role = afterContext.substring(0, afterContext.indexOf("/"));
-        }
-        return role;
-    }
 }
